requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.18 #for correct data presentation
tabulate==0.8.6
